Refactor: Time related stuff
