+++ src/engine.hpp
@@ -7,7 +7,6 @@
 #include "chess.hpp"
 #include "constants.hpp"
 #include "search.hpp"
-#include "time_manager.hpp"
 #include "tt.hpp"
 
 class Engine {
@@ -18,9 +17,6 @@ class Engine {
   void printBoard();
 
   void initilizeEngine();
-
-  void setSearchLimits(int wtime, int btime, int winc, int binc, int movestogo,
-                       int movetime);
 
   void makeMove(std::string move);
 
@@ -39,7 +35,6 @@ class Engine {
  private:
   chess::Board board;
   TranspositionTable tt_helper;
-  TimeManager time_manager;
   Search search;
 
   std::thread searchThread;
